+++ src/main/java/org/torproject/descriptor/impl/RelayNetworkStatusConsensusImpl.java
@@ -54,8 +54,9 @@ public class RelayNetworkStatusConsensusImpl extends NetworkStatusImpl
     Set<String> atMostOnceKeywords = new HashSet<>(Arrays.asList((
         "client-versions,server-versions,recommended-client-protocols,"
         + "recommended-relay-protocols,required-client-protocols,"
-        + "required-relay-protocols,params,directory-footer,"
-        + "bandwidth-weights").split(",")));
+        + "required-relay-protocols,params,shared-rand-previous-value,"
+        + "shared-rand-current-value,directory-footer,bandwidth-weights")
+        .split(",")));
     this.checkAtMostOnceKeywords(atMostOnceKeywords);
     this.checkFirstKeyword("network-status-version");
     this.clearParsedKeywords();
@@ -146,6 +147,12 @@ public class RelayNetworkStatusConsensusImpl extends NetworkStatusImpl
           break;
         case "params":
           this.parseParamsLine(line, parts);
+          break;
+        case "shared-rand-previous-value":
+          this.parseSharedRandPreviousValueLine(line, parts);
+          break;
+        case "shared-rand-current-value":
+          this.parseSharedRandCurrentValueLine(line, parts);
           break;
         default:
           if (this.failUnrecognizedDescriptorLines) {
@@ -350,6 +357,36 @@ public class RelayNetworkStatusConsensusImpl extends NetworkStatusImpl
         parts, 1, "=");
   }
 
+  private void parseSharedRandPreviousValueLine(String line, String[] parts)
+      throws DescriptorParseException {
+    if (parts.length != 3) {
+      throw new DescriptorParseException("Illegal line '" + line
+          + "' in vote.");
+    }
+    try {
+      this.sharedRandPreviousNumReveals = Integer.parseInt(parts[1]);
+    } catch (NumberFormatException e) {
+      throw new DescriptorParseException("Illegal line '" + line
+          + "' in vote.");
+    }
+    this.sharedRandPreviousValue = parts[2];
+  }
+
+  private void parseSharedRandCurrentValueLine(String line, String[] parts)
+      throws DescriptorParseException {
+    if (parts.length != 3) {
+      throw new DescriptorParseException("Illegal line '" + line
+          + "' in vote.");
+    }
+    try {
+      this.sharedRandCurrentNumReveals = Integer.parseInt(parts[1]);
+    } catch (NumberFormatException e) {
+      throw new DescriptorParseException("Illegal line '" + line
+          + "' in vote.");
+    }
+    this.sharedRandCurrentValue = parts[2];
+  }
+
   private void parseBandwidthWeightsLine(String line, String[] parts)
       throws DescriptorParseException {
     this.bandwidthWeights = ParseHelper.parseKeyValueIntegerPairs(line,
@@ -484,6 +521,34 @@ public class RelayNetworkStatusConsensusImpl extends NetworkStatusImpl
   public SortedMap<String, Integer> getConsensusParams() {
     return this.consensusParams == null ? null
         : new TreeMap<>(this.consensusParams);
+  }
+
+  private int sharedRandPreviousNumReveals = -1;
+
+  @Override
+  public int getSharedRandPreviousNumReveals() {
+    return this.sharedRandPreviousNumReveals;
+  }
+
+  private String sharedRandPreviousValue = null;
+
+  @Override
+  public String getSharedRandPreviousValue() {
+    return this.sharedRandPreviousValue;
+  }
+
+  private int sharedRandCurrentNumReveals = -1;
+
+  @Override
+  public int getSharedRandCurrentNumReveals() {
+    return this.sharedRandCurrentNumReveals;
+  }
+
+  private String sharedRandCurrentValue = null;
+
+  @Override
+  public String getSharedRandCurrentValue() {
+    return this.sharedRandCurrentValue;
   }
 
   private SortedMap<String, Integer> bandwidthWeights;